+++ stripe_terminal/CHANGELOG.md
@@ -1,4 +1,6 @@
+
 # 4.0.0
+- fix(android): execute reader disconnect after hot restart in main thread
 - build!: Bumped [Android](https://github.com/stripe/stripe-terminal-android/blob/master/CHANGELOG.md#410---2024-11-18)
   sdk version to `4.1.0` and [IOS](https://github.com/stripe/stripe-terminal-android/blob/master/CHANGELOG.md#410---2024-11-18)
   sdk version to `4.1.0`. Please watch the official CHANGELOG.md to know the breaking changes